Convert MultiSelectFiltered to tsx
